add sacrifice system (untested)

+++ contracts/src/components/occupied.cairo
@@ -4,5 +4,5 @@
 struct Occupied {
     #[key]
     entity_id: u128,
-    until: u128,
+    until: u32,
 }
